Refactor ShowCandidate component and add new features
This commit refactors the ShowCandidate component by renaming the `$candidat` property to `$candidate` for consistency. It also adds a new property `$activeTab` to track the active tab in the component.

Additionally, this commit introduces new methods `getAcademieInfo()`, `getEtablissementInfo()`, `getSpecialitesInfo()`, and `getFormePassageInfo()` to retrieve specific information related to the candidate.

Lastly, the `setActiveTab()` method is added to allow changing the active tab dynamically.

These changes improve the functionality and maintainability of the ShowCandidate component.